==> 01. date_format_correction.ipynb <==
# Date format correction

import pandas as pd
import numpy as np
import datetime as dt
from datetime import datetime
from datetime import timedelta
from dateutil.relativedelta import relativedelta

We need the date in the dataset about transfers in the format %y-%m-%d so that we can compare it with the data of each match we will find in the dataset of each player

d_tf = pd.read_csv('data_5.csv')

for i in range(0,d_tf.shape[0]):
    data_split = d_tf.date[i].split('-')
    month = data_split[0].replace('gen','01').replace('feb','02').replace('mar','03').replace('giu','06').replace('lug','07').replace('ago','08').replace('set','09').replace('mag','05').replace('dic','12')
    d_tf.date[i]= d_tf["year"].map(str)[i] + '-' + month + '-' + data_split[1]
d_tf.to_csv("data_6.csv", sep=",", index=False)
