==> 02. Join with wyscout players.ipynb <==
# Joins with players.json dataset

import pandas as pd
import numpy as np

d_tf = pd.read_csv('data_6.csv')
dp = pd.read_csv('df_players_1.csv')

dp = dp.drop(['middleName', 'name', 'currentNationalTeamId','currentTeamId', 'alpha2code_y','alpha2code_x'], axis=1) 

dp['name_final'] = np.nan
def lastnameExtract(x):
    return x.split(' ')[-1]

dp['name_final'] = dp.shortName.apply(lastnameExtract)

d_tf1 = d_tf.join(dp.set_index('name_final'), on='name_final')

d_tf1.shape 

d_tf1.to_csv('data_7.csv', header=True, sep=',')

d_tf.shape

d_tf1.isna().sum()

# 127 NA needs to be imput and 501 needs to be selected by hand!
