==> 03. country selling and buying germany 1617.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv("data_to_add.csv",encoding = "utf8")
df.head()

dt1 = pd.read_csv("dati_teams_frometonations.csv", encoding='utf8')

dteams = pd.read_csv("data_4_update.csv",encoding = "utf8")
dteams = dteams.rename(columns={'team_from':'team_to'})

dteams.head()

dt2 = dt1.join(dteams.set_index('team_to'), on='team_to')
 

dt2.to_csv('dati_teams_frometonations.csv', header=True, sep=',')
