==> 04. Corrections.ipynb <==
# Corrections

Correct by hand: lugo is Spain, not Italy. Brentford is England, not Germany

Correct by hand:
Danilo 
G. Pezzella (German e Giuseppe) 
N. Maksimovic
D. Brooks e J. Brooks

import pandas as pd
import numpy as np
import sys 
import csv
import unicodedata
import collections

df = pd.read_csv('data_17b.csv', encoding='utf8')

def remove_accents(input_str):
    nkfd_form = unicodedata.normalize('NFKD', input_str)
    return u"".join([c for c in nkfd_form if not unicodedata.combining(c)])

def remove_special (input_str):
    new = input_str.replace("'", "").replace("æ", "ae").replace("Ð", "D").replace("ð", "d").replace("Ø", "O").replace("ø", "o").replace("Æ", "AE").replace("œ", "oe").replace('ς','c').replace('đ','d')
    return new

df['name_final'] = df.shortName.apply(remove_accents).apply(remove_special)

df.shape

tot = [item for item, count in collections.Counter(df.name_final).items() if count > 1]

Di questi 58 devo escludere 2 prestiti e correggerne 4 A MANO:


# exclude loans
df = df[(df.name_final != 'Bruno Fernandes') | (df.year != 2016)]
df = df[(df.name_final != 'G. Castrovilli') | (df.price != 1.2)]

df.shape

# Team's data correction

df = df.replace('Beri 1908','Bari 1908').replace('Ostende','Oostende').replace('Norwich','Norwich City').replace('Nuremberg','Nurnberg').replace('AZ Alkmaar','AZ').replace('Braga','Sporting Braga').replace('Extremadura','Extremadura UD')

# remove hednesford town because of lack of data

# remove Sturm graz that is in austria
df = df[df.team_from != 'Sturm Graz']
df.shape

# remove hednesford town because we have no data about team's performance
df = df[(df.team_from != 'Hednesford Town')]
df.shape

df.to_csv('data_10b_by_hand.csv', header=True, sep=',')
