==> 05. european_ranking.ipynb <==
import pandas as pd
import numpy as np

# open the 3 datasets
df_cl = pd.read_excel("UEFAperTeam.xlsx",encoding = "utf8")
d_tf = pd.read_csv("data_10b_by_hand.csv",encoding = "utf8")

d_tf.team_from = d_tf.team_from.replace('Bolton','Bolton Wanderers').replace('Charlton','Charlton Athletic').replace('Leeds','Leeds United').replace('QPR','Queens Park Rangers').replace('Pacos Ferreira','Pacos de Ferreira').replace('1860 Munich','1860 Munchen').replace('Hull','Hull City').replace('Derby','Derby County').replace('Bari','Beri 1908').replace('Latina Calcio','Latina').replace('Alcoron','Alcorcon').replace('FC Barcelona B','Barcelona II').replace('Leicester','Leicester City').replace('CD Numancia','Numancia').replace('NEC Nijmegen','NEC').replace('CD Lugo','Lugo')
d_tf.team_to = d_tf.team_to.replace('Bolton','Bolton Wanderers').replace('Charlton','Charlton Athletic').replace('Leeds','Leeds United').replace('QPR','Queens Park Rangers').replace('Pacos Ferreira','Pacos de Ferreira').replace('1860 Munich','1860 Munchen').replace('Hull','Hull City').replace('Derby','Derby County').replace('Bari','Beri 1908').replace('Latina Calcio','Latina').replace('Alcoron','Alcorcon').replace('FC Barcelona B','Barcelona II').replace('CD Numancia','Numancia').replace('NEC Nijmegen','NEC').replace('CD Lugo','Lugo')

df_cl.rename(columns={'Team':'team_from'}, 
                 inplace=True)

# join the 2 datasets on the team selling the player
d_tf2 = d_tf.join(df_cl.set_index('team_from'), on='team_from')

d_tf2.head()

d_tf2['europeanPerf_team_from'] = np.nan

def perfEurYearBefore(df):
        if df['season']=='16/17':
            return df['15/16']
        elif df['season']=='17/18':
            return df['16/17']
        elif df['season']=='18/19':
            return df['17/18']
        else:
            return np.nan

d_tf2['europeanPerf_team_from'] = d_tf2.apply(perfEurYearBefore, axis = 1).fillna(0)
d_tf2 = d_tf2.drop(['country','14/15','15/16','16/17','17/18','18/19','Nation'], axis=1)

# bad programming:
#d_tf2['europeanPerf_team_from'] = np.nan

#for i in range(0, d_tf2.shape[0]):
#    if d_tf2['season'][i]=='16/17':
#        d_tf2['europeanPerf_team_from'][i] = d_tf2['15/16'][i]
#    elif d_tf2['season'][i]=='17/18':
#        d_tf2['europeanPerf_team_from'][i] = d_tf2['16/17'][i]
#    elif d_tf2['season'][i]=='18/19':
#        d_tf2['europeanPerf_team_from'][i] = d_tf2['17/18'][i]

#d_tf2['europeanPerf_team_from'] = d_tf2.europeanPerf_team_from.fillna(0)

df_cl.rename(columns={'team_from':'team_to'}, 
                 inplace=True)

# join the 2 datasets on the team of arrival
d_tf2 = d_tf2.join(df_cl.set_index('team_to'), on='team_to')

d_tf2['europeanPerf_team_to'] = np.nan
d_tf2['europeanPerf_team_to'] = d_tf2.apply(perfEurYearBefore, axis = 1).fillna(0)
d_tf2 = d_tf2.drop(['country','14/15','15/16','16/17','17/18','18/19','Nation'], axis=1)

d_tf2.head(10)

d_tf2.to_csv("data_11.csv", header=True, sep=',')
