==> 06. national_ranking.ipynb <==
# Add national performance to the transfer's dataset

import pandas as pd
import numpy as np

d_tf  = pd.read_csv('data_11.csv', encoding = "utf8")

dpn = pd.read_csv('data_12.csv', encoding="utf8")

# recode teams' names
dpn.Team = dpn.Team.replace('Schalke','Schalke 04').replace('AC Milan','Milan').replace('ACF Fiorentina','Fiorentina').replace('Atalanta BC','Atalanta').replace( 'AS Monaco','Monaco').replace( 'AS Roma','Roma').replace( 'AS Saint0Étienne', 'Saint-Etienne').replace( 'Bayer 04 Leverkusen','Bayer Leverkusen').replace( 'Beşiktaş JK','Besiktas').replace( 'Club Atlético de Madrid','Atletico Madrid').replace( 'Estoril Praia','Estoril').replace( 'Fenerbahçe SK','Fenerbahce').replace( 'Galatasaray AŞ', 'Galatasaray').replace( 'Girondins de Bordeaux','Bordeaux').replace( 'Heracles Almelo','Heracles').replace( 'Hull City AFC','Hull').replace( 'Hertha BSC Berlin', 'Hertha BSC').replace( 'Internazionale Milano','Internazionale').replace( 'KAA Gent','Gent').replace( 'KRC Genk','Genk').replace( 'KV Oostende','Oostende').replace( 'LOSC Lille','Lille').replace( 'OGC Nice','Nice').replace( 'Olympique de Marseille','Olympique Marseille').replace( 'PSV Eindhoven','PSV').replace( 'RC Celta de Vigo','Celta de Vigo').replace( 'Real Betis Balompié','Real Betis').replace( 'Real Sociedad de Fútbol','Real Sociedad')
dpn.Team = dpn.Team.str.replace('Gijon','Sporting Gijon').replace('Augsburg ','Augsburg').replace('Bayern Munich','Bayern Munchen').replace('Dortmund','Borussia Dortmund').replace('Hannover','Hannover 96').replace('B. Monchengladbach','Borussia Mgladbach').replace('FC Koln','Koln').replace('Werder Bremen ','Werder Bremen').replace('Betis','Real Betis').replace('Dusseldorf ','Dusseldorf').replace('Ferreira','Pacos de Ferreira').replace('Stuttgart ','Stuttgart').replace('Bayer Leverkusen ','Bayer Leverkusen').replace('Hoffenheim ','Hoffenheim').replace('Eintracht Frankfurt ','Eintracht Frankfurt').replace('Hamburger SV','Hamburger SV').replace('Guimaraes','Vitoria Guimaraes').replace('Hull','Hull City').replace('Eupen','KAS Eupen').replace('Spal','SPAL').replace('Club Brugge KV','Club Brugge').replace('Zwolle','PEC Zwolle').replace('Nijmegen','NEC').replace('Oostende','Ostende').replace('Huddersfield','Huddersfield Town').replace('Rio Ave FC','Rio Ave').replace('Ingolstadt','FC Ingolstadt').replace('Ingolstadt 04','FC Ingolstadt').replace('Hull','Hull City').replace('Nurnberg','Nuremberg').replace('Norwich City','Norwich').replace().replace('FC Twente','Twente').replace('B. Monchengladbach','Borussia Mgladbach').replace('Sporting Lisbon','Sporting CP').replace('Lyon','Olympique Lyonnais').replace('Stade Rennais','Rennes').replace('KRC Genk','Genk').replace('Celta Vigo','Celta de Vigo').replace('Huddersfield','Huddersfield Town').replace('Sporting Braga','Braga').replace('Wolves','Wolverhampton').replace('FC Utrecht','Utrecht').replace('Volfsburg', 'Wolfsburg').replace('Vfl Wolfsburg','Wolfsburg').replace('Bor. Mgladbach', 'Borussia Mgladbach').replace('Monchengladbach', 'Borussia Mgladbach').replace(' B. Monchengladbach', 'Borussia Mgladbach').replace('Leverkusen', 'Bayer Leverkusen').replace('Celta Vigo','Celta de Vigo').replace('Celta','Celta de Vigo').replace('Chievo Verona', 'Chievo').replace('Bor. Dortmund', 'Borussia Dortmund').replace('Dep. La Coruna', 'Deportivo La Coruna').replace('Inter', 'Internazionale').replace('Inter Milan','Internazionale').replace('Herrha Berlin', 'Hertha BSC').replace('Hertha Berlin', 'Hertha BSC').replace('Bayern Munich','Bayern Munchen').replace('AC Milan', 'Milan').replace('AS Monaco','Monaco').replace('Paris SG', 'PSG').replace('AS Roma', 'Roma').replace('FC Barcelona', 'Barcelona').replace('1FC Köln', 'Koln').replace('1FC Köln', 'Koln').replace('FC Koln', 'Koln').replace('West Ham', 'West Ham United').replace('West Brom', 'West Bromwich Albion').replace('West Brom', 'West Bromwich Albion').replace('West Bromwich', 'West Bromwich Albion').replace('W. Bremen', 'Werder Bremen').replace('Tottenham','Tottenham Hotspur').replace('Swansea', 'Swansea City').replace('Manchester Utd','Manchester United').replace('Man City','Manchester City').replace('Verona', 'Hellas Verona').replace('Eint. Frankfurt','Eintracht Frankfurt').replace('Frankfurt', 'Eintracht Frankfurt').replace('Leicester', 'Leicester City').replace('Olympique Lyon','Olympique Lyonnais').replace('Lyon Marseille','Olympique Lyonnais').replace('Alaves', 'Deportivo Alaves').replace('Stoke', 'Stoke City').replace('Newcastle', 'Newcastle United').replace('Schalke','Schalke 04').replace('VfB Stuttgart','Stuttgart').replace('Stutgart','Stuttgart').replace('Samdoria', 'Sampdoria').replace('Sevilla Atletico','Sevilla').replace('Hannover','Hannover 96').replace('FC Metz', 'Metz').replace('Mainz', 'Mainz 05').replace('FC Porto', 'Porto').replace('Levante UD','Levante').replace('Marseille','Olympique Marseille').replace('Brighton', 'Brighton Hove Albion').replace('OGC Nice','Nice').replace('St. Etienne', 'Saint-Etienne').replace('Athletic Bilbao','Athletic Club').replace('Vit. Guimaraes', 'Vitoria Guimaraes').replace('KAA Gent','Gent').replace('Hamburg','Hamburger SV').replace('PSV Eindhoven','PSV').replace('Sporting','Sporting CP')

dpn = dpn.drop(['Unnamed: 0'], axis=1)

dpn.rename(columns={'Team':'team_from'}, 
                 inplace=True)

# join the 2 datasets on the team selling the player
d_tf2 = d_tf.join(dpn.set_index('team_from'), on='team_from')

d_tf2 = d_tf2.drop(['Unnamed: 0'], axis=1)

# Modify columns: 1516 1617 1718
I always consider the mean of the results obtained in 15 days of championship. So we have to modify the columns relative to the whole year, according to the number of days in each league. 38 for spa, eng, ita, fra. 34 for ger, ola, port and 30 for belgium

# 0 al posto di na: se c'è 0 sono in serie B
d_tf2['1516'] = d_tf2['1516'].fillna(0)
d_tf2['1617'] = d_tf2['1617'].fillna(0)
d_tf2['1718'] = d_tf2['1718'].fillna(0)

def meanOn15Days1516(df):
    if df['country_selling'] == 'GB' or df['country_selling'] == 'IT' or df['country_selling'] == 'FR' or df['country_selling'] == 'ES':
        return round(df['1516']*15/38)
    elif df['country_selling'] == 'NL' or df['country_selling'] == 'PR' or df['country_selling'] == 'DE':
        return round(df['1516']*15/34)
    elif df['country_selling'] == 'BE':
        return round(df['1516']*15/30)
    else:
        return np.nan

d_tf2['1516'] = d_tf2.apply(meanOn15Days1516, axis = 1)

def meanOn15Days1617(df):
    if df['country_selling'] == 'GB' or df['country_selling'] == 'IT' or df['country_selling'] == 'FR' or df['country_selling'] == 'ES':
        return round(df['1617']*15/38)
    elif df['country_selling'] == 'NL' or df['country_selling'] == 'PR' or df['country_selling'] == 'DE':
        return round(df['1617']*15/34)
    elif df['country_selling'] == 'BE':
        return round(df['1617']*15/30)
    else:
        return np.nan
    
d_tf2['1617'] = d_tf2.apply(meanOn15Days1617, axis = 1)

def meanOn15Days1718(df):
    if df['country_selling'] == 'GB' or df['country_selling'] == 'IT' or df['country_selling'] == 'FR' or df['country_selling'] == 'ES':
        return round(df['1718']*15/38)
    elif df['country_selling'] == 'NL' or df['country_selling'] == 'PR' or df['country_selling'] == 'DE':
        return round(df['1718']*15/34)
    elif df['country_selling'] == 'BE':
        return round(df['1718']*15/30)
    else:
        return np.nan

d_tf2['1718'] = d_tf2.apply(meanOn15Days1718, axis = 1)

d_tf2['nationalPerf_team_from'] = np.nan

def perfYearBefore(df):
    if df['year'] == 2016 and df['wintersummer']=='s':
        return df['1516']
    elif df['year'] == 2017 and df['wintersummer']=='s':
        return df['1617']
    elif df['year'] == 2018 and df['wintersummer']=='s':
        return df['1718']
    elif df['year'] == 2017 and df['wintersummer']=='w':
        return df['1617_half']
    elif df['year'] == 2018 and df['wintersummer']=='w':
        return df['1718_half']
    elif df['year'] == 2019 and df['wintersummer']=='w':
        return df['1819_half']
    else:
        return np.nan

d_tf2['nationalPerf_team_from'] = d_tf2.apply(perfYearBefore, axis = 1)

# drop columns with the rankings of all years
d_tf2 = d_tf2.drop(['1516','1617_half','1617','1718_half','1718','1819_half'], axis=1)

# Dummy creation: 0 if it was not in the 1st division when the player was sold

d_tf2['nationalPerf_team_from'] = d_tf2.nationalPerf_team_from.fillna(0)

d_tf2['dummyAB'] = np.nan

def dummyAB(df):
    if df['nationalPerf_team_from'] == 0:
        return 0
    else:
        return 1
    
# 0 = not A, 1 = A
d_tf2['dummyAB'] = d_tf2.apply(dummyAB, axis = 1)

d_tf2.head()

d_tf2.to_csv('data_13.csv', sep=',', header=True)
