==> 07. money_team_to.ipynb <==
# Dummies about the earnings of the buying teams

import pandas as pd
import numpy as np

d_tf  = pd.read_csv('data_13.csv', encoding = "utf8").drop(['Unnamed: 0'], axis=1)

dd = pd.read_excel('data_13_deloitte.xlsx', encoding="utf8")

dd['team_to'] = dd.index
dd.head(20)

# Dummy creation: previous year earnings in the first5

d_tf2 = d_tf.join(dd.set_index('team_to'), on='team_to')

d_tf2['y1516'] = d_tf2['y1516'].fillna(0)
d_tf2['y1617'] = d_tf2['y1617'].fillna(0)
d_tf2['y1718'] = d_tf2['y1718'].fillna(0)

num1516 = sorted(list(d_tf2['y1516'].fillna(0).unique()), reverse=True)[4]
num1617 = sorted(list(d_tf2['y1617'].fillna(0).unique()), reverse=True)[4]
num1718 = sorted(list(d_tf2['y1718'].fillna(0).unique()), reverse=True)[4]

d_tf2['dummyfirst5'] = 0
d_tf2.loc[(d_tf2['season']=='16/17') & (d_tf2['y1516'] >= num1516), 'dummyfirst5'] = 1
d_tf2.loc[(d_tf2['season']=='17/18') & (d_tf2['y1617'] >= num1617), 'dummyfirst5'] = 1
d_tf2.loc[(d_tf2['season']=='18/19') & (d_tf2['y1718'] >= num1718), 'dummyfirst5'] = 1

# first 10

num1516 = sorted(list(d_tf2['y1516'].fillna(0).unique()), reverse=True)[9]
num1617 = sorted(list(d_tf2['y1617'].fillna(0).unique()), reverse=True)[9]
num1718 = sorted(list(d_tf2['y1718'].fillna(0).unique()), reverse=True)[9]

d_tf2['dummyfirst10'] = 0
d_tf2.loc[(d_tf2['season']=='16/17') & (d_tf2['y1516'] >= num1516), 'dummyfirst10'] = 1
d_tf2.loc[(d_tf2['season']=='17/18') & (d_tf2['y1617'] >= num1617), 'dummyfirst10'] = 1
d_tf2.loc[(d_tf2['season']=='18/19') & (d_tf2['y1718'] >= num1718), 'dummyfirst10'] = 1

# first 20

num1516 = sorted(list(d_tf2['y1516'].fillna(0).unique()), reverse=True)[18]
num1617 = sorted(list(d_tf2['y1617'].fillna(0).unique()), reverse=True)[18]
num1718 = sorted(list(d_tf2['y1718'].fillna(0).unique()), reverse=True)[18]

d_tf2['dummyfirst20'] = 0
d_tf2.loc[(d_tf2['season']=='16/17') & (d_tf2['y1516'] >= num1516), 'dummyfirst20'] = 1
d_tf2.loc[(d_tf2['season']=='17/18') & (d_tf2['y1617'] >= num1617), 'dummyfirst20'] = 1
d_tf2.loc[(d_tf2['season']=='18/19') & (d_tf2['y1718'] >= num1718), 'dummyfirst20'] = 1

d_tf2.head()

# Dummy creation: PSG & Man City

d_tf2['dummySheikh'] = np.nan

def dummySheikh(df):
    if df['team_to'] == 'PSG' or df['team_to'] == 'Manchester City':
        return 1
    else:
        return 0
    
# 1 = sheikh, 0 =  no sheikh

d_tf2['dummySheikh'] = d_tf2.apply(dummySheikh, axis = 1)

d_tf2 = d_tf2.drop(['y1415','y1617','y1718','y1516'],axis=1)

d_tf2.to_csv('data_14.csv',sep=',', header=True)
