==> 08. age_when_transferred_data16.ipynb <==
# Age when transferred

import pandas as pd
import numpy as np
import datetime as dt
from datetime import datetime
from datetime import timedelta
from dateutil.relativedelta import relativedelta

df = pd.read_csv('data_14.csv')

df['Age']=np.nan

df['date'] = pd.to_datetime(df['date'], format="%d/%m/%Y")
df['birthDate'] = pd.to_datetime(df['birthDate'])

df['Age'] = df['date'] - df['birthDate']
df['Age']=round(df['Age']/np.timedelta64(1,'Y'))

df.to_csv('data_15.csv', header=True, sep=',')
