==> 10. number_players_bought_and_sold.ipynb <==
# Add number per nations and per teams bought and sold

import pandas as pd
import numpy as np
from numpy import *
import scipy.stats as scp

d_team_player = pd.read_csv('data_16.csv', encoding = "utf8")

# number of export players
num_exported = d_team_player.groupby('team_from').count().iloc[:,0]

dataExport = pd.DataFrame(data = num_exported)

dataExport.rename(columns = {'Unnamed: 0':'number_export_players'}, inplace=True)
dataExport.index.name = 'team_from'
dataExport.reset_index(inplace=True)

d_team_player = d_team_player.join(dataExport.set_index('team_from'), on='team_from')
where_are_NaNs = isnan(d_team_player.number_export_players)
d_team_player.number_export_players[where_are_NaNs] = 0
d_team_player.number_export_players = round(d_team_player.number_export_players)

d_team_player = d_team_player.drop(columns=['Unnamed: 0','Unnamed: 0.1'], axis=1)

d_team_player = d_team_player.drop(columns=['Unnamed: 0.1.1', 'Unnamed: 0.1.1.1'], axis=1)

# number of import players
num_imported = d_team_player.groupby('team_to').count().iloc[:,0]

dataImport = pd.DataFrame(data = num_imported)

dataImport.rename(columns = {'date':'number_import_players'}, inplace=True)
dataImport.index.name = 'team_to'
dataImport.reset_index(inplace=True)

d_team_player = d_team_player.join(dataImport.set_index('team_to'), on='team_to')
where_are_NaNs = isnan(d_team_player.number_import_players)
d_team_player.number_import_players[where_are_NaNs] = 0
d_team_player.number_import_players = round(d_team_player.number_import_players)

# absolute number of players sold from a country

num_sold = d_team_player.groupby('country_selling').count().iloc[:,0]

num_sold = pd.DataFrame(data = num_sold)
num_sold.rename(columns = {'date':'abs_number_per_nation_sold'}, inplace=True)
num_sold.index.name = 'country_selling'
num_sold.reset_index(inplace=True)

d_team_player = d_team_player.join(num_sold.set_index('country_selling'), on='country_selling')

# absolute number of players bought from a country
num_bought = d_team_player.groupby('country_buying').count().iloc[:,0]

num_bought = pd.DataFrame(data = num_bought)
num_bought.rename(columns = {'date':'abs_number_per_nation_bought'}, inplace=True)
num_bought.index.name = 'country_buying'
num_bought.reset_index(inplace=True)

d_team_player = d_team_player.join(num_bought.set_index('country_buying'), on='country_buying')


d_team_player.to_csv('data_17.csv', header=True, sep=',')
