==> 12. Concat files.ipynb <==
# How to concatenate excel files?

import pandas as pd

# concat 4 files with the same name with exeption to the team

team_low= 'Fiorentina'
df = pd.read_excel("./prova_teams/Team Stats " + team_low + ".xlsx",encoding = "utf8")
df1 = pd.read_excel("./prova_teams/Team Stats " + team_low + " (1).xlsx",encoding = "utf8")
df2 = pd.read_excel("./prova_teams/Team Stats " + team_low + " (2).xlsx",encoding = "utf8")
df3 = pd.read_excel("./prova_teams/Team Stats " + team_low + " (3).xlsx",encoding = "utf8")

df1.head()

df = df.drop(df.index[0:2]) # remove two white lines
df1 = df1.drop(df1.index[0:2])
df2 = df2.drop(df2.index[0:2])
df3 = df3.drop(df3.index[0:2])

df_tot = pd.concat([df,df1, df2,df3])
