==> 14. Create dummies and liquidity, drop and rename columns.ipynb <==
# Create dummies

import pandas as pd
import numpy as np
import sys 
import csv
import unicodedata
import collections
import scipy

df = pd.read_csv('data_19_noGoalkeeper.csv', encoding = "utf8")

df.shape

# more than 1 transfer in 3 years
tot = [item for item, count in collections.Counter(df.name_final).items() if count > 1]
df['more1t'] = np.nan
df['more1t'] = [1 if x in tot else 0 for x in df['name_final']]

# dummy more than 18 years
df['d18'] = np.nan
df['d18'] = [1 if x <19 else 0 for x in df['Age']]

df.shape

# liquidity
df1 = df.groupby(["wintersummer", "year"]).size().reset_index(name="liq")
df2 = pd.merge(df, df1,  how='outer', left_on=['wintersummer','year'], right_on = ['wintersummer','year'])

df2.shape

# national performance == 0 devo dividere B-C-D
bcd = pd.read_csv('bcd.csv', encoding='utf8').drop(columns=['weight','height','Unnamed: 0','dummyA'], axis=1)

bcd.head()

pd.to_datetime(df2.date)
date_format = '%Y-%m-%d'
df2['date'] = pd.to_datetime(df2['date'])
df2.date

pd.to_datetime(bcd.date)
date_format = '%d/%m/%Y'
bcd['date'] = pd.to_datetime(bcd['date'])

df3 = pd.merge(df2, bcd,  how='outer', left_on=['team_from','date'], right_on = ['team_from','date'])

df3['div'] = df3['div'].fillna('a')
df3['dA'] = np.nan
df3['dB'] = np.nan
df3['dC'] = np.nan

df3['dA'] = [1 if x == 'a' else 0 for x in df3['div']]
df3['dB'] = [1 if x == 'b' else 0 for x in df3['div']]
df3['dC'] = [1 if x == 'c' else 0 for x in df3['div']]

df3.shape

df3 = df3.drop_duplicates() # non so perché si creano ma li tolgo

df3.shape

### rimuovo ultime righe che sono un errore che viene dal merge

# Days from open of the transfer window of the buyer team -> 0 if negative

df3.days_from_openBuying[df3.days_from_openBuying < 0] = 0

# Drop columns

df3 = df3.drop(columns=['Unnamed: 0','year','birthDate','nationalPerf_team_from','shortName','dummyAB'], axis=1)

df4 = df3.drop(columns=['country_name_selling','name_final','country_name_buying','date','city_selling','city_buying','PassportCountry'], axis=1)

df3.to_csv('data_graphs.csv', header=True, sep=',')

# Rename columns

df4.rename(columns={'team_from':'TF','team_to':'TT','wintersummer':'ws', 'country_selling':'countryS','country_buying':'countryB','birthCountry':'birth','europeanPerf_team_from':'euroTS','europeanPerf_team_to':'euroTB','dummyfirst5':'d5','dummyfirst10':'d10','dummyfirst20':'d20','dummySheikh':'dSheikh','days_from_openBuying':'openB','europeanPerf_country_from':'euroCS','europeanPerf_country_to':'euroCB','number_export_players':'noExp','number_import_players':'noImp','abs_number_per_nation_sold':'noCS','abs_number_per_nation_bought':'noCB','number_matches':'noMatch','number_matches1Year':'noMatch1Y','number_matches6Months':'noMatch6','perc_s_actions':'actionS','goal_per_match':'goals','assist_per_match':'assist','perc_shots_target':'shotS','passes_acc':'passA','lpasses_acc':'lpassA','crosses_acc':'crossA','perc_dribbles':'dribbS','perc_duels':'duelS','perc_aerialduels':'aerDuelS','interceptions':'interc','rec_opposite':'recOpp','rec_tot':'recT','los_own':'losOwn','los_tot':'losTot','minutes1Year':'min1Y','perc_s_actions1Year':'actionS1Y','goal_per_match1Year':'goal1Y','assist_per_match1Year':'assist1Y','perc_shots_target1Year':'shotS1Y','xg1Year':'xg1Y','passes_acc1Year':'passA1Y','lpasses_acc1Year':'lpassA1Y','crosses_acc1Year':'crossA1Y','perc_dribbles1Year':'dribbS1Y','perc_duels1Year':'duelS1Y','perc_aerialduels1Year':'aerDuelS1Y','interceptions1Year':'interc1Y','rec_opposite1Year':'recOpp1Y','rec_tot1Year':'recT1Y','los_own1Year':'losOwn1Y','los_tot1Year':'losT1Y','yellow1Year':'yell1Y','red1Year':'red1Y','minutes1Year':'min1Y','minutes6Months':'min6','perc_s_actions6Months':'actionS6','goal_per_match6Months':'goal6','assist_per_match6Months':'assist6','perc_shots_target6Months':'shotS6','xg6Months':'xg6','passes_acc6Months':'passA6','lpasses_acc6Months':'lpassA6','crosses_acc6Months':'crossA6','perc_dribbles6Months':'dribbS6','perc_duels6Months':'duelS6','perc_aerialduels6Months':'aerDuelS6','interceptions6Months':'interc6','rec_opposite6Months':'recOpp6','rec_tot6Months':'recT6','los_own6Months':'losOwn6','los_tot6Months':'losT6','yellow6Months':'yell6','red6Months':'red6','shots_per_match':'shotT','perc_shots_target_team':'shotST','xg_team':'xgT','goals_per_match_team':'goalT','passes_per_match':'passT','passes_accuracy':'passAT','perc_possession_team':'possT','los_per_match_team':'losT','low_losses_team':'lowLosT','rec_per_match_team':'recT','high_recoveries_team':'highRecT','challenges_per_match_team':'challT','perc_challenges_won_team':'challST','total_goals_opponents':'goalOppT','total_goals':'goalAbsT','shots_per_match1Year':'shotsT1Y','perc_shots_target_team1Year':'shotST1Y','xg_team1Year':'xgT1Y','goals_per_match_team1Year':'goalT1Y','passes_per_match1Year':'passT1Y','passes_accuracy1Year':'passAT1Y','perc_possession_team1Year':'possT1Y','los_per_match_team1Year':'losT1Y','low_losses_team1Year':'lowLosT1Y','rec_per_match_team1Year':'recT1Y','high_recoveries_team1Year':'highRecT1Y','challenges_per_match_team1Year':'challT1Y','perc_challenges_won_team1Year':'challST1Y','total_goals_opponents1Year':'goalOppT1Y','total_goals1Year':'goalAbsT1Y','shots_per_match6Months':'shotT6','perc_shots_target_team6Months':'shotST6','xg_team6Months':'xgT6','goals_per_match_team6Months':'goalT6','passes_per_match6Months':'passT6','passes_accuracy6Months':'passAT6','perc_possession_team6Months':'possT6','los_per_match_team6Months':'losT6','low_losses_team6Months':'lowLosT6','rec_per_match_team6Months':'recT6','high_recoveries_team6Months':'highRecT6','challenges_per_match_team6Months':'challT6','perc_challenges_won_team6Months':'challST6','total_goals_opponents6Months':'goalOppT6','total_goals6Months':'goalAbsT6'}, 
                 inplace=True)

pd.set_option('display.max_columns', 500)
df3.head()

nomicol = pd.DataFrame(df4.columns)

df4.to_csv('data_models.csv', header=True, sep=',')

df4.shape

sum(df4.noCB.unique())-767
